+++ src/render/draw_fill_extrusion.ts
@@ -13,7 +13,7 @@ import type FillExtrusionStyleLayer from '../style/style_layer/fill_extrusion_st
 import type FillExtrusionBucket from '../data/bucket/fill_extrusion_bucket';
 import type {OverscaledTileID} from '../source/tile_id';
 
-import findPatternPositions from './draw_fill_helper';
+import updatePatternPositionsInProgram from './update_pattern_positions_in_program';
 
 export default draw;
 
@@ -78,7 +78,7 @@ function drawExtrusionTiles(
             programConfiguration.updatePaintBuffers(crossfade);
         }
 
-        findPatternPositions(fillPropertyName, constantPattern, tile, layer, programConfiguration);
+        updatePatternPositionsInProgram(programConfiguration, fillPropertyName, constantPattern, tile, layer);
 
         const matrix = painter.translatePosMatrix(
             coord.posMatrix,
